+++ src/lib/converter/plugins/CommentPlugin.ts
@@ -267,11 +267,13 @@ export class CommentPlugin extends ConverterComponent {
      */
     private onResolve(context: Context, reflection: Reflection) {
         if (reflection.comment) {
-            reflection.label = extractLabelTag(reflection.comment);
-            if (reflection.label && !/[A-Z_][A-Z0-9_]/.test(reflection.label)) {
+            if (
+                reflection.comment.label &&
+                !/[A-Z_][A-Z0-9_]/.test(reflection.comment.label)
+            ) {
                 context.logger.warn(
                     `The label "${
-                        reflection.label
+                        reflection.comment.label
                     }" for ${reflection.getFriendlyFullName()} cannot be referenced with a declaration reference. ` +
                         `Labels may only contain A-Z, 0-9, and _, and may not start with a number.`
                 );
@@ -483,15 +485,6 @@ function moveNestedParamTags(comment: Comment, parameter: ParameterReflection) {
     parameter.type?.visit(visitor);
 }
 
-function extractLabelTag(comment: Comment): string | undefined {
-    const index = comment.summary.findIndex(
-        (part) => part.kind === "inline-tag" && part.tag === "@label"
-    );
-
-    if (index !== -1) {
-        return comment.summary.splice(index, 1)[0].text;
-    }
-}
 function mergeSeeTags(comment: Comment) {
     const see = comment.getTags("@see");