+++ src/lib/models/reflections/kind.ts
@@ -28,44 +28,78 @@ export enum ReflectionKind {
     Reference = 0x800000,
 }
 
-/** @hidden */
 export namespace ReflectionKind {
+    // Would be nice to derive this...
+    export type KindString =
+        | "Project"
+        | "Module"
+        | "Namespace"
+        | "Enum"
+        | "EnumMember"
+        | "Variable"
+        | "Function"
+        | "Class"
+        | "Interface"
+        | "Constructor"
+        | "Property"
+        | "Method"
+        | "CallSignature"
+        | "IndexSignature"
+        | "ConstructorSignature"
+        | "Parameter"
+        | "TypeLiteral"
+        | "TypeParameter"
+        | "Accessor"
+        | "GetSignature"
+        | "SetSignature"
+        | "ObjectLiteral"
+        | "TypeAlias"
+        | "Reference";
+
     export const All = ReflectionKind.Reference * 2 - 1;
 
+    /** @internal */
     export const ClassOrInterface =
         ReflectionKind.Class | ReflectionKind.Interface;
+    /** @internal */
     export const VariableOrProperty =
         ReflectionKind.Variable | ReflectionKind.Property;
+    /** @internal */
     export const FunctionOrMethod =
         ReflectionKind.Function | ReflectionKind.Method;
+    /** @internal */
     export const ClassMember =
         ReflectionKind.Accessor |
         ReflectionKind.Constructor |
         ReflectionKind.Method |
         ReflectionKind.Property;
+    /** @internal */
     export const SomeSignature =
         ReflectionKind.CallSignature |
         ReflectionKind.IndexSignature |
         ReflectionKind.ConstructorSignature |
         ReflectionKind.GetSignature |
         ReflectionKind.SetSignature;
+    /** @internal */
     export const SomeModule = ReflectionKind.Namespace | ReflectionKind.Module;
+    /** @internal */
     export const SomeType =
         ReflectionKind.Interface |
         ReflectionKind.TypeLiteral |
         ReflectionKind.TypeParameter |
         ReflectionKind.TypeAlias;
+    /** @internal */
     export const SomeValue =
         ReflectionKind.Variable |
         ReflectionKind.Function |
         ReflectionKind.ObjectLiteral;
-
+    /** @internal */
     export const SomeMember =
         ReflectionKind.EnumMember |
         ReflectionKind.Property |
         ReflectionKind.Method |
         ReflectionKind.Accessor;
-
+    /** @internal */
     export const SomeExport =
         ReflectionKind.Module |
         ReflectionKind.Namespace |
@@ -76,7 +110,7 @@ export namespace ReflectionKind {
         ReflectionKind.Interface |
         ReflectionKind.TypeAlias |
         ReflectionKind.Reference;
-
+    /** @internal */
     export const ExportContainer =
         ReflectionKind.SomeModule | ReflectionKind.Project;
 
@@ -93,4 +127,40 @@ export namespace ReflectionKind {
         ReflectionKind.Constructor |
         ReflectionKind.Function |
         ReflectionKind.Method;
+
+    const SINGULARS = {
+        [ReflectionKind.Enum]: "Enumeration",
+        [ReflectionKind.EnumMember]: "Enumeration Member",
+    };
+
+    const PLURALS = {
+        [ReflectionKind.Class]: "Classes",
+        [ReflectionKind.Property]: "Properties",
+        [ReflectionKind.Enum]: "Enumerations",
+        [ReflectionKind.EnumMember]: "Enumeration Members",
+        [ReflectionKind.TypeAlias]: "Type Aliases",
+    };
+
+    export function singularString(kind: ReflectionKind): string {
+        if (kind in SINGULARS) {
+            return SINGULARS[kind as keyof typeof SINGULARS];
+        } else {
+            return getKindString(kind);
+        }
+    }
+
+    export function pluralString(kind: ReflectionKind): string {
+        if (kind in PLURALS) {
+            return PLURALS[kind as keyof typeof PLURALS];
+        } else {
+            return getKindString(kind) + "s";
+        }
+    }
+}
+
+function getKindString(kind: ReflectionKind): string {
+    return ReflectionKind[kind].replace(
+        /(.)([A-Z])/g,
+        (_m, a, b) => a + " " + b.toLowerCase()
+    );
 }