+++ src/lib/output/themes/default/partials/type.tsx
@@ -61,7 +61,11 @@ function getNamespacedPath(reflection: Reflection): Reflection[] {
 }
 function renderUniquePath(context: DefaultThemeRenderContext, reflection: Reflection): JSX.Element {
     return join(<span class="tsd-signature-symbol">.</span>, getUniquePath(reflection), (item) => (
-        <a href={context.urlTo(item)} class="tsd-signature-type" data-tsd-kind={item.kindString}>
+        <a
+            href={context.urlTo(item)}
+            class="tsd-signature-type"
+            data-tsd-kind={ReflectionKind.singularString(item.kind)}
+        >
             {item.name}
         </a>
     ));
@@ -243,7 +247,7 @@ const typeRenderers: {
             if (reflection.kindOf(ReflectionKind.TypeParameter)) {
                 // Don't generate a link will always point to this page, but do set the kind.
                 name = (
-                    <span class="tsd-signature-type" data-tsd-kind={reflection.kindString}>
+                    <span class="tsd-signature-type" data-tsd-kind={ReflectionKind.singularString(reflection.kind)}>
                         {reflection.name}
                     </span>
                 );